pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "facet"
version = "0.1.0"
description = "Facet — 信息提取 → 认知模型 → 量化投资"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "jinozaki" }]

dependencies = [
    # ── Anchor (数据提取) ──
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Database
    "sqlmodel>=0.0.21",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29.0",
    "alembic>=1.14.0",
    "aiosqlite>=0.20.0",
    # HTTP & Scraping
    "httpx>=0.27.0",
    "feedparser>=6.0.11",
    # Social media APIs
    "tweepy>=4.14.0",
    "youtube-transcript-api>=1.2.0",
    "pytubefix>=8.0.0",
    "av>=12.0.0",
    # AI
    "anthropic>=0.40.0",
    # Scheduling
    "apscheduler>=3.10.4",
    # Config
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.5.0",
    # Financial data
    "yfinance>=0.2.40",
    "akshare>=1.14.0",
    # SEC EDGAR
    "edgartools>=3.0.0",
    # Macro data sources
    "fredapi>=0.5.1",
    "pandas>=2.0.0",
    "pandas-datareader>=0.10.0",
    "wbgapi>=1.0.12",
    "dbnomics>=0.9.3",
    # Utilities
    "tenacity>=9.0.0",
    "loguru>=0.7.2",
    "python-dateutil>=2.9.0",
    "pyyaml>=6.0",
    # CLI
    "click>=8.0",
    # ── Polaris (认知模型) ──
    "numpy>=1.26",
    # ── Axion (量化执行) ──
    # (no extra deps beyond pandas/numpy already listed)
]

[project.optional-dependencies]
pg = ["psycopg2-binary>=2.9"]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.4.0",
]

[project.scripts]
anchor = "anchor.cli:main"
polaris = "polaris.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/anchor", "src/polaris", "src/axion"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 120
